agent-driven: Move get_binary_prop to ucd_helper.rs
